test: 예외사항 추가
